+++ tests/mod.rs
@@ -236,15 +236,32 @@ fn enable_day_alarm_ok() {
 }
 
 #[test]
-fn enable_second_interrupt_ok() {
+fn enable_second_interrupt_pulsed_ok() {
     let expectations = [
         I2cTransaction::read(PCF8523_I2C_ADDRESS, [0b0].to_vec()),
         i2c_reg_read(PCF8523_CONTROL_1, 0b0010_0011),
         i2c_reg_write(PCF8523_CONTROL_1, 0b0010_0111),
+        i2c_reg_read(PCF8523_TMR_CLKOUT_CTRL, 0b0010_0011),
+        i2c_reg_write(PCF8523_TMR_CLKOUT_CTRL, 0b1010_0011),
     ];
     let mut i2c = I2cMock::new(&expectations);
     let mut driver = Pcf8523::new(&mut i2c).unwrap();
-    driver.enable_second_interrupt().unwrap();
+    driver.enable_second_interrupt(true).unwrap();
+    i2c.done();
+}
+
+#[test]
+fn enable_second_interrupt_permanent_ok() {
+    let expectations = [
+        I2cTransaction::read(PCF8523_I2C_ADDRESS, [0b0].to_vec()),
+        i2c_reg_read(PCF8523_CONTROL_1, 0b0010_0011),
+        i2c_reg_write(PCF8523_CONTROL_1, 0b0010_0111),
+        i2c_reg_read(PCF8523_TMR_CLKOUT_CTRL, 0b1010_0011),
+        i2c_reg_write(PCF8523_TMR_CLKOUT_CTRL, 0b0010_0011),
+    ];
+    let mut i2c = I2cMock::new(&expectations);
+    let mut driver = Pcf8523::new(&mut i2c).unwrap();
+    driver.enable_second_interrupt(false).unwrap();
     i2c.done();
 }